Add Assistant as a collapsible sidebar
