+++ pages/video/Video.module.scss
@@ -1,37 +1,13 @@
-@use '../../styles/colors' as c;
-@use 'sass:color' as color;
-
 .outerWrapper {
-  padding: 1rem;
-}
-
-
-.textAndChatWrapper {
-  display: flex;
-  height: 80%;
-  width: 80%;
-  border-radius: 14px;
-  background-color: c.$bg-black;
-  box-shadow: 0 0 6px 1px color.change(c.$bg-black, $alpha: 0.6);
-  color: white;
-  overflow: hidden;
-  > div {
-    &:first-of-type {
-      border-right: 1px solid black;
-    }
-    flex: 1 0 0;
-  }
-}
-
-.chatWrapper {
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
-  overflow: hidden;
+  height: 100vh;
+  // padding: 1rem;
 }
 
-.chatBubbles {
-  flex: 1 1 0;
-  padding: 1rem;
-  overflow-y: auto;
+.mainSection {
+  display: flex;
+  width: 100%;
+  height: 100%;
 }
+
